backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlmodel==0.0.14
aiosqlite==0.20.0
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-docx==1.1.0
jinja2==3.1.3
google-genai>=1.0.0
markdown==3.6
pytest==8.0.2
pytest-asyncio==0.23.5
anyio==4.3.0
fpdf2>=2.7.0
